+++ CHANGELOG.md
@@ -4,6 +4,9 @@ CHANGELOG
 0.3.0
 -----
 
+* exceptions thrown during instantiating value objects are now caught and mapped back to the form as transformation
+  failures
+* [BC BREAK] the `ValueObjectTransformer` requires an `ExceptionHandlerRegistry` instance
 * [BC BREAK] the first argument's type of the `ExceptionHandlerInterface::getError()` method has been changed from
   `FormInterface` to `FormConfigInterface`
 * [BC BREAK] the return type of `ExceptionHandlerInterface::getError()` has been changed to `SensioLabs\RichModelForms\ExceptionHandling\Error`